Split run_on_whole_repo functions to separate file and mark as internal
