+++ src/pre_commit_terraform/terraform_fmt.py
@@ -12,7 +12,7 @@ from subprocess import PIPE
 from subprocess import run
 from typing import Final
 
-from pre_commit_terraform import common
+from pre_commit_terraform import _common as common
 from pre_commit_terraform._types import ReturnCodeType
 from pre_commit_terraform.logger import setup_logging
